CVE-2016-0830: fix syntax with fix crashes with lots of discovered LE devices
fix syntax of patch CVE-2016-0830   ANDROID-26071376
Fix crashes with lots of discovered LE devices

When loads of devices are discovered a config file which is too large
can be written out, which causes the BT daemon to crash on startup.

This limits the number of config entries for unpaired devices which
are initialized, and prevents a large number from being saved to the
filesystem.